Defer numba guvectorize compilation for popgen functions
